+++ packages/react-components/src/BlockchainAddress/BlockchainAddress.types.ts
@@ -3,10 +3,9 @@ import {
     BlockchainLinksTypesOverrides,
 } from "@peersyst/react-components-core";
 import { OverridableStringUnion } from "@peersyst/react-types";
-import { BlockchainAddressEllipsis } from "@peersyst/react-utils";
-import { TypographyProps } from "../Typography";
+import { HashProps } from "src/Hash";
 
-export interface BlockchainAddressProps extends Omit<TypographyProps, "singleLine" | "children"> {
+export type BlockchainAddressProps = Omit<HashProps, "url" | "hash"> & {
     /**
      * Blockchain address
      */
@@ -15,23 +14,7 @@ export interface BlockchainAddressProps extends Omit<TypographyProps, "singleLin
      * Blockchain address type
      */
     type: OverridableStringUnion<BlockchainLinksTypes, BlockchainLinksTypesOverrides>;
-    /**
-     * Address' length in characters
-     */
-    length?: number | "auto" | "complete";
-    /**
-     * If ellipsis should be in the middle or end
-     */
-    ellipsis?: BlockchainAddressEllipsis;
-    /**
-     * If address should break
-     */
-    break?: boolean;
-    /**
-     * Gap between the address and copy button
-     */
-    gap?: number;
-}
+};
 
 export interface BlockchainAddressRootProps {
     autoLength: boolean;