+++ packages/react-components/src/Hash/Hash.styles.ts
@@ -1,0 +1,24 @@
+import styled, { css } from "styled-components";
+import { HashRootProps, HashTextProps, HashLinkProps } from "./Hash.types";
+import { Row } from "../Row";
+import { Typography } from "../Typography";
+
+export const HashRoot = styled(Row).attrs({
+    alignItems: "center",
+})<HashRootProps>`
+    width: ${(p) => p.autoLength && "fit-content"};
+    max-width: 100%;
+`;
+
+export const HashLink = styled.a<HashLinkProps>(
+    ({ url }) => css`
+        text-decoration: none;
+        &:hover {
+            text-decoration: ${url ? "underline" : "none"};
+        }
+    `,
+);
+
+export const HashText = styled(Typography)<HashTextProps>`
+    word-break: ${(p) => p.break && "break-word"};
+`;